Remove file only if there is other to move in place

+++ move_files_to_data.sh
@@ -9,8 +9,10 @@ mv $HOME/Downloads/nubank*.csv ./data/
 # Spliwise csv exports from mobile app
 mv $HOME/Downloads/Splitwise*.csv ./data/
 
-# Remove current splitwise exports
-rm ./data/mozi-e-eu*.csv
+# Remove current splitwise exports only if there is another to move
+if [ -f $HOME/Downloads/mozi-e-eu*.csv ]; then
+  rm ./data/mozi-e-eu*.csv
+fi
 
 # Spliwise csv exports from site
 mv $HOME/Downloads/mozi-e-eu*.csv ./data/
